# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: AI 刷题 App API
  version: 0.1.0
  description: |
    首版接口契约（阶段 1：可靠刷题闭环 + 学习档案 + 管理端内容维护）。
    - 前缀 /api/v1；JSON 字段 camelCase，枚举值小写 snake_case。
    - 时间为 UTC RFC 3339；列表使用游标分页 ?cursor=&limit=（默认 20，最大 100）。
    - 认证：HttpOnly Cookie（session）；写请求校验 Origin。
    - 错误统一结构：{"error": {"code", "message", "requestId"}}；前端分支只依赖稳定 code。
    - 核心边界：提交前接口永不返回答案、解析或判分字段；answer_keys 仅判分与管理端可读。
servers:
  - url: /api/v1
...
paths:
...
  /knowledge-points:
    get:
      tags: [learning]
      summary: 知识点列表（含个人统计；仅已发布）
      parameters:
        - { name: levelId, in: query, schema: { type: string } }
        - { name: subjectId, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string }, description: 名称模糊搜索 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  knowledgePoints:
                    type: array
                    items: { $ref: '#/components/schemas/KnowledgePointItem' }
  /knowledge-points/{id}:
    get:
      tags: [learning]
      summary: 知识点详情（正文仅已发布可见）与个人统计
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/KnowledgePointDetail' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /wrong-items:
    get:
      tags: [learning]
      summary: 错题本（每题取最近一次错误作答）
      parameters: [{ name: knowledgePointId, in: query, schema: { type: string } }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  wrongItems:
                    type: array
                    items: { $ref: '#/components/schemas/WrongItem' }
  /issue-reports:
    post:
      tags: [learning]
      summary: 提交题目问题反馈（后端补全题目版本与判分上下文）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [practiceItemId, targetType]
              properties:
                practiceItemId: { type: string }
                targetType: { type: string, enum: [stem, answer, explanation, classification, ai_grading] }
                description: { type: string }
      responses:
        '201': { description: 已提交 }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /admin/issue-reports:
    get:
      tags: [admin]
      summary: 举报列表
      parameters: [{ name: status, in: query, schema: { type: string, enum: [open, resolved, dismissed] } }]
      responses: { '200': { description: OK } }
  /admin/issue-reports/{id}:
    patch:
      tags: [admin]
      summary: 处理举报（resolved / dismissed + 备注）
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: OK }
        '409': { description: issue_already_handled }
components:
  responses:
...
    NotFound:
      description: not_found（不泄露资源是否存在）
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: { type: object, nullable: true }
...
    AnswerValue:
      oneOf:
        - type: object
          properties: { optionIds: { type: array, items: { type: string } } }
        - type: object
          properties: { text: { type: string } }
      nullable: true
...
    KnowledgePointItem:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        levelId: { type: string }
        levelCode: { type: string }
        subjectId: { type: string }
        subjectName: { type: string }
        parentId: { type: string, nullable: true }
        questionCount: { type: integer }
        stats:
          nullable: true
          type: object
          properties:
            confirmedAnswered: { type: integer }
            confirmedCorrect: { type: integer }
            recentAnswered: { type: integer }
            recentCorrect: { type: integer }
            aiAnswered: { type: integer }
            aiCorrect: { type: integer }
            consecutiveWrong: { type: integer }
            lastPracticedAt: { type: string, nullable: true }
    KnowledgePointDetail:
      allOf:
        - $ref: '#/components/schemas/KnowledgePointItem'
        - type: object
          properties:
            description: { type: string }
            commonMistakes: { type: string }
            examples: { type: string }
            status: { type: string }
    WrongItem:
      type: object
      properties:
        itemId: { type: string }
        sessionId: { type: string }
        questionId: { type: string }
        position: { type: integer }
        type: { type: string }
        stem: { type: string }
        options: { type: array, items: { type: object }, nullable: true }
        material: { type: object, nullable: true }
        knowledgePoints: { type: array, items: { type: object } }
        gradingStatus: { type: string }
        answerAuthority: { type: string, nullable: true }
        userAnswer: { $ref: '#/components/schemas/AnswerValue' }
        correctAnswer: { $ref: '#/components/schemas/AnswerValue' }
        explanation: { type: object, nullable: true }
...